requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
neo4j==5.16.0
mcp
mcp[cli]
openai
# openai-whisper
pdfminer.six
python-multipart==0.0.7
# yt-dlp